requirements.txt
setuptools~=75.7.0
python-taiga~=1.3.0
